feat: Add member_of method to Users struct
This commit adds a new method `member_of` to the `Users` struct in the `users.rs` file. The `member_of` method retrieves a list of users who are members of a specified group. It filters the users based on the `member` field, which contains a HashSet of group names.

This change improves the functionality of the `Users` struct by allowing users to easily retrieve a list of members for a given group.